Make storeManager => BulkStoreCacher submodule
