+++ runtime/bus.go
@@ -131,6 +131,10 @@ func (m *bus) GetIBCHost() modules.IBCHostModule {
 	return getModuleFromRegistry[modules.IBCHostModule](m, modules.IBCHostModuleName)
 }
 
+func (m *bus) GetBulkStoreCacher() modules.BulkStoreCacher {
+	return getModuleFromRegistry[modules.BulkStoreCacher](m, modules.BulkStoreCacherModuleName)
+}
+
 // getModuleFromRegistry is a helper function to get a module from the registry that handles errors and casting via generics
 func getModuleFromRegistry[T modules.Submodule](m *bus, moduleName string) T {
 	mod, err := m.modulesRegistry.GetModule(moduleName)
